fix: filters + configurations
