+++ apps/artemis-web/components/ongoing-hijack-table/ongoing-hijack-table.tsx
@@ -445,11 +445,10 @@ const OngoingHijackTableComponent = (props: any): ReactElement => {
             key={i}
             value={option.text}
             onClick={() => onSizePerPageChange(option.page)}
-            className={`btn ${
-              currSizePerPage === `${option.page}`
-                ? 'btn-secondary'
-                : 'btn-warning'
-            }`}
+            className={`btn ${currSizePerPage === `${option.page}`
+              ? 'btn-secondary'
+              : 'btn-warning'
+              }`}
           >
             {option.text}
           </option>
@@ -558,7 +557,13 @@ const OngoingHijackTableComponent = (props: any): ReactElement => {
         return (
           <>
             <div className="header-filter">
-              <SizePerPageDropdownStandalone {...paginationProps} />
+              <div style={{ float: "left" }}>
+                <SizePerPageDropdownStandalone {...paginationProps} />
+              </div>
+              <div style={{ float: "right" }}>
+                <PaginationTotalStandalone {...paginationProps} />
+                <PaginationListStandalone {...paginationProps} />
+              </div>
             </div>
             <BootstrapTable
               remote
@@ -590,8 +595,6 @@ const OngoingHijackTableComponent = (props: any): ReactElement => {
               {...toolkitprops.baseProps}
               {...paginationTableProps}
             />
-            <PaginationTotalStandalone {...paginationProps} />
-            <PaginationListStandalone {...paginationProps} />
           </>
         );
       }}
