+++ apps/artemis-web/next.config.js
@@ -7,6 +7,9 @@ module.exports = {
   },
   poweredByHeader: false,
   i18n: { locales: ['en'], defaultLocale: 'en' },
+  future: {
+    webpack5: true,
+  },
   async headers() {
     return [
       {
