+++ package.json
@@ -47,7 +47,7 @@
     "jsonwebtoken": "^8.5.0",
     "mongodb": "^3.5.2",
     "nanoid": "^3.1.12",
-    "next": "^10.0.0",
+    "next": "^10.2.0",
     "next-connect": "^0.9.0",
     "next-csrf": "^0.1.1",
     "next-session": "^3.3.2",
@@ -100,6 +100,7 @@
     "@types/react-dom": "17.0.0",
     "@typescript-eslint/eslint-plugin": "4.15.0",
     "@typescript-eslint/parser": "4.15.0",
+    "@webpack-cli/generators": "^2.0.0",
     "babel-jest": "26.6.3",
     "babel-loader": "^8.0.0-beta.0",
     "css-loader": "^5.0.2",
@@ -129,9 +130,9 @@
     "ts-node": "~9.1.1",
     "typescript": "~4.1.3",
     "url-loader": "^4.1.1",
-    "webpack": "^5.21.2",
-    "webpack-cli": "^4.5.0",
-    "webpack-dev-server": "^3.7.0"
+    "webpack": "^5.36.1",
+    "webpack-cli": "^4.6.0",
+    "webpack-dev-server": "^3.11.2"
   },
   "husky": {
     "hooks": {
